修正　警告（DWORD から　floatへの変換）Direct3Dクラスのプロジェクション行列作成の関数　アスペクト比の計算部分　キャストを使用することで対応
branch : DX9Base/dev_sub2

+++ DirectX9_Base/Source/Direct3D/Direct3D.cpp
@@ -468,7 +468,7 @@ void Direct3D::LoadMeshX(MeshX& mesh,TCHAR* path)
 
 		mesh.ppTextureFileNames = new TCHAR*[mesh.numMaterials];
 
-		for (int i = 0; i < mesh.numMaterials; i++)
+		for (unsigned int i = 0; i < mesh.numMaterials; i++)
 		{
 			mesh.ppTextureFileNames[i] = nullptr;
 		}
@@ -637,8 +637,8 @@ void Direct3D::SetupRrojectionMatrix()
 {
 	RECT client;
 	GetClientRect(hWnd, &client);
-	float w = client.right - client.left;
-	float h = client.bottom - client.top;
+	float w =static_cast<float> (client.right - client.left);
+	float h =static_cast<float> (client.bottom - client.top);
 
 	D3DXMATRIXA16 matProj;
